README.md
Depression among students of different age groups is on the increase daily. This analysis aims to fine-tune the causes of this student depression and offer a possible solution to the menace. The data set used is from Kaggle, and Microsoft Excel was used for the analysis using pivot tables and charts.